+++ api/src/main/java/jakarta/security/jacc/PolicyConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2020 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.jacc;
+package jakarta.security.jacc;
 
 import java.security.Permission;
 import java.security.PermissionCollection;
@@ -148,7 +148,7 @@ import java.security.PermissionCollection;
  * @see java.security.Permission
  * @see java.security.PermissionCollection
  * @see PolicyContextException
- * @see javax.security.jacc.PolicyConfigurationFactory
+ * @see jakarta.security.jacc.PolicyConfigurationFactory
  *
  * @author Ron Monzillo
  * @author Gary Ellison
