+++ pom.xml
@@ -28,9 +28,9 @@
 
     <groupId>org.eclipse.ee4j.jacc</groupId>
     <artifactId>jacc-parent</artifactId>
+    <version>2.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
-    <version>1.5.0-SNAPSHOT</version>
-    
+
     <name>Jakarta Authorization Parent</name>
     <description>
         Jakarta Authorization defines a low-level SPI for authorization modules, which are repositories of permissions 
@@ -39,28 +39,6 @@
         these permissions.
     </description>
     <url>https://github.com/eclipse-ee4j/jacc</url>
-
-    <scm>
-        <connection>scm:git:ssh://git@github.com/eclipse-ee4j/jacc.git</connection>
-        <developerConnection>scm:git:ssh://git@github.com/eclipse-ee4j/jacc.git</developerConnection>
-        <url>https://github.com/eclipse-ee4j/jacc</url>
-    </scm>
-
-    <properties>
-        <non.final>false</non.final>
-        <extension.name>javax.security.jacc</extension.name>
-        <spec.version>1.5</spec.version>
-        <findbugs.version>2.4.0</findbugs.version>
-        <findbugs.skip>false</findbugs.skip>
-        <findbugs.exclude>exclude-common.xml</findbugs.exclude>
-        <findbugs.threshold>Low</findbugs.threshold>
-    </properties>
-
-    <modules>
-        <module>api</module>
-        <module>spec</module>
-    </modules>
-
     <licenses>
         <license>
             <name>EPL 2.0</name>
@@ -74,4 +52,40 @@
         </license>
     </licenses>
 
+    <modules>
+        <module>api</module>
+        <module>spec</module>
+    </modules>
+
+    <scm>
+        <connection>scm:git:ssh://git@github.com/eclipse-ee4j/jacc.git</connection>
+        <developerConnection>scm:git:ssh://git@github.com/eclipse-ee4j/jacc.git</developerConnection>
+        <url>https://github.com/eclipse-ee4j/jacc</url>
+    </scm>
+
+    <build>
+        <plugins>
+            <!-- Sets minimal Maven version to 3.5.4 -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-enforcer-plugin</artifactId>
+                <version>3.0.0-M2</version>
+                <executions>
+                    <execution>
+                        <id>enforce-maven</id>
+                        <goals>
+                            <goal>enforce</goal>
+                        </goals>
+                        <configuration>
+                            <rules>
+                                <requireMavenVersion>
+                                    <version>3.5.4</version>
+                                </requireMavenVersion>
+                            </rules>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
